requirements.txt
apache-airflow[postgres,crypto]==2.1.2
pandas==1.3.1